+++ src/String.ts
@@ -375,10 +375,10 @@ export const dropLeftWhile = (f: Predicate<string>): Endomorphism<string> =>
  *
  * @example
  * import { dropRightWhile } from 'fp-ts-std/String';
- * import { elemFlipped } from 'fp-ts-std/Array';
+ * import { elemV } from 'fp-ts-std/Array';
  * import { eqString } from 'fp-ts/Eq';
  *
- * const isVowel = elemFlipped(eqString)(['a', 'e', 'i', 'o', 'u']);
+ * const isVowel = elemV(eqString)(['a', 'e', 'i', 'o', 'u']);
  * const dropRightVowels = dropRightWhile(isVowel);
  *
  * assert.deepStrictEqual(dropRightVowels('hellooo'), 'hell');