+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to semantic versioning.
 - Add `memptyWhen` and `memptyUnless` functions to the `Monoid` module.
 - Add `memptyWhen` and `memptyUnless` functions to the `Option` module.
 - Remove support for variadic/non-unary functions from `flip`.
+- Rename the "Flipped" suffix wherever present to "V".
 
 ## 0.12.0 (2021-12-01)